+++ lib/action_policy/lookup_chain.rb
@@ -22,7 +22,7 @@ module ActionPolicy
         attr_reader :store
 
         def fetch(namespace, policy)
-          return yield unless LookupChain.namespace_cache_enabled
+          return yield unless LookupChain.namespace_cache_enabled?
           return store[namespace][policy] if store[namespace].key?(policy)
           store[namespace][policy] ||= yield
         end
@@ -37,6 +37,8 @@ module ActionPolicy
 
     class << self
       attr_accessor :chain, :namespace_cache_enabled
+
+      alias namespace_cache_enabled? namespace_cache_enabled
 
       def call(record, **opts)
         chain.each do |probe|
@@ -76,8 +78,10 @@ module ActionPolicy
       end
     end
 
-    # Enable namespace cache by default
-    self.namespace_cache_enabled = true
+    # Enable namespace cache by default or
+    # if RACK_ENV provided and equal to "production"
+    self.namespace_cache_enabled =
+      !ENV["RACK_ENV"].nil? ? ENV["RACK_ENV"] == "production" : true
 
     # By self `policy_class` method
     INSTANCE_POLICY_CLASS = ->(record, _) {